Fixed CNN hyperparams to numericals

+++ src/CNN_train_test.py
@@ -127,15 +127,16 @@ input_size_full = train_x_std_all.shape[1]
 
 # train/test CNN1
 if args.cnn_type == "cnn_1":
-    hidden_size_conv  = float(config["CNN_hyperparam_1"]["hidden_size_conv"])
-    kernel_size       = float(config["CNN_hyperparam_1"]["kernel_size"])
-    padding           = float(config["CNN_hyperparam_1"]["padding"])
-    hidden_size_full  = float(config["CNN_hyperparam_1"]["hidden_size_full"])
-    dropout_full      = float(config["CNN_hyperparam_1"]["dropout_full"])
-    hidden_size_combo = float(config["CNN_hyperparam_1"]["hidden_size_combo"])
-    dropout_combo     = float(config["CNN_hyperparam_1"]["dropout_combo"])
-    lr                = float(config["CNN_hyperparam_1"]["lr"])
-    weight_decay      = float(config["CNN_hyperparam_1"]["weight_decay"])
+    hidden_size_conv  = int(config["CNN_hyperparam_1"]["hidden_size_conv"]) 
+    kernel_size       = int(config["CNN_hyperparam_1"]["kernel_size"])
+    padding           = int(config["CNN_hyperparam_1"]["padding"])
+    hidden_size_full  = int(config["CNN_hyperparam_1"]["hidden_size_full"])
+    dropout_full      = np.float64(config["CNN_hyperparam_1"]["dropout_full"])
+    hidden_size_combo = int(config["CNN_hyperparam_1"]["hidden_size_combo"])
+    dropout_combo     = np.float64(config["CNN_hyperparam_1"]["dropout_combo"])
+    lr                = np.float64(config["CNN_hyperparam_1"]["lr"])
+    weight_decay      = np.float64(config["CNN_hyperparam_1"]["weight_decay"])
+
 
     # Loss function
     mse_loss = torch.nn.MSELoss(size_average=True)
@@ -179,16 +180,15 @@ if args.cnn_type == "cnn_1":
 
 # train/test CNN2
 else:
-    hidden_size_conv  = float(config["CNN_hyperparam_2"]["hidden_size_conv"])
-    kernel_size       = float(config["CNN_hyperparam_2"]["kernel_size"])
-    padding           = float(config["CNN_hyperparam_2"]["padding"])
-    hidden_size_full  = float(config["CNN_hyperparam_2"]["hidden_size_full"])
-    dropout_full      = float(config["CNN_hyperparam_2"]["dropout_full"])
-    hidden_size2_full = float(config["CNN_hyperparam_2"]["hidden_size2_full"])
-    dropout2_full     = float(config["CNN_hyperparam_2"]["dropout2_full"])
-    lr                = float(config["CNN_hyperparam_2"]["lr"])
-    weight_decay      = float(config["CNN_hyperparam_2"]["weight_decay"])
-
+    hidden_size_conv  = int(config["CNN_hyperparam_2"]["hidden_size_conv"]) 
+    kernel_size       = int(config["CNN_hyperparam_2"]["kernel_size"])
+    padding           = int(config["CNN_hyperparam_2"]["padding"])
+    hidden_size_full  = int(config["CNN_hyperparam_2"]["hidden_size_full"])
+    dropout_full      = np.float64(config["CNN_hyperparam_2"]["dropout_full"])
+    hidden_size2_full = int(config["CNN_hyperparam_2"]["hidden_size2_full"])
+    dropout2_full     = np.float64(config["CNN_hyperparam_2"]["dropout2_full"])
+    lr                = np.float64(config["CNN_hyperparam_2"]["lr"])
+    weight_decay      = np.float64(config["CNN_hyperparam_2"]["weight_decay"])
 
     # Loss function
     mse_loss = torch.nn.MSELoss(size_average=True)
